BAC-45: producst by category slug
