+++ app/schemas/category.py
@@ -1,21 +1,6 @@
-from typing import Optional
+from typing import TYPE_CHECKING, List, Optional
 
 from pydantic import Field
 
 from .orm import OrmBaseModel
-
-
-class CategoryIn(OrmBaseModel):
-    parent_id: Optional[int] = Field(
-        None, description="The ID of the parent category.", example=42)
-    image_url: str = Field(..., description="The URL of the category image.",
-                           example="https://allthatsinteresting.com/wordpress/wp-content/uploads/2012/06/iconic-photos-1950-einstein.jpg")
-    name: str = Field(..., description="The name of the category.",
-                      example="Eggs & Dairy")
-    description: str = Field(..., description="The description of the category.",
-                             example="Our eggs and dairy products are locally source and delivered rotten!")
-
-class CategoryOut(CategoryIn):
-    id: int = Field(..., description="The ID of the category.", example=420)
-    slug: str = Field(..., description="The slug of the category.",
-                      example="eggs-dairy")
+from .product import CategoryIn, CategoryOut
